Refactor class FileToRelativePath as a pre-requisite to fix unleash-scm-provider-git
- Deprecate com.itemis.maven.plugins.unleash.util.FileToRelativePath
  -> this is moved to com.itemis.maven.plugins.unleash.scm.utils.FileToRelativePath
- Others:
  - followup this refactoring
