+++ plugin/src/test/java/com/itemis/maven/plugins/unleash/util/functions/FileToRelativePathTest.java
@@ -9,6 +9,7 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 
+import com.itemis.maven.plugins.unleash.scm.utils.FileToRelativePath;
 import com.tngtech.java.junit.dataprovider.DataProvider;
 import com.tngtech.java.junit.dataprovider.DataProviderRunner;
